% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNmap.R
\docType{data}
\name{CNmap}
\alias{CNmap}
\title{National Map of China}
\format{
## `CNmap`
A sf tibble with 1 rows and 4 columns:
\describe{
  \item{ID}{`100000`}
  \item{name}{Country name in Chinese.}
  \item{level}{`country` level.}
  \item{geometry}{list column of simple feature.}
}
}
\source{
<https://github.com/Civitasv/DataV_GeoJSON>
}
\usage{
CNmap
}
\description{
A national map of China, which consider all regions of China as a whole.
}
\keyword{datasets}
