% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNmap_county.R
\docType{data}
\name{CNmap_county}
\alias{CNmap_county}
\title{County Level Map of China}
\format{
## `CNmap_county`
A sf tibble with 2847 rows and 4 columns:
\describe{
  \item{ID}{Geocoding of each county.}
  \item{name}{County name in Chinese.}
  \item{level}{`district` level.}
  \item{geometry}{list column of simple feature.}
}
}
\source{
<https://github.com/Civitasv/DataV_GeoJSON>
}
\usage{
CNmap_county
}
\description{
County level map of China.
}
\keyword{datasets}
