% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNmap_province.R
\docType{data}
\name{CNmap_province}
\alias{CNmap_province}
\title{Province Level Map of China}
\format{
## `CNmap_province`
A sf tibble with 34 rows and 4 columns:
\describe{
  \item{ID}{Geocoding of each Province.}
  \item{name}{Province name in Chinese.}
  \item{level}{`province` level.}
  \item{geometry}{list column of simple feature.}
}
}
\source{
<https://github.com/Civitasv/DataV_GeoJSON>
}
\usage{
CNmap_province
}
\description{
Province level map of China.
}
\keyword{datasets}
