% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{extract_province}
\alias{extract_province}
\alias{extract_city}
\alias{extract_county}
\title{Subset Map Data}
\usage{
extract_province(map, ...)

extract_city(map, ...)

extract_county(map, ...)
}
\arguments{
\item{map}{map data in `CNmap` package.}

\item{...}{other parameters passing to `*_id()` function.}
}
\value{
a sf object.
}
\description{
A set of functions to subset of CNmap.
}
\author{
Hou Yun
}
