% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-poi.R
\name{get_poi}
\alias{get_poi}
\title{Get POI}
\usage{
get_poi(query, region = 340, ..., api = NULL)
}
\arguments{
\item{query}{key words of query.}

\item{region}{region of query.}

\item{...}{other parameters passing to API.}

\item{api}{Baidu Map API key.}
}
\value{
a data frame.
}
\description{
Location search service.
}
