% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{province_id}
\alias{province_id}
\alias{city_id}
\alias{county_id}
\title{Extract ID}
\usage{
province_id(..., output = c("num", "chr"), grep = TRUE, na.rm = TRUE)

city_id(
  ...,
  province = NULL,
  output = c("num", "chr"),
  grep = TRUE,
  na.rm = TRUE
)

county_id(
  ...,
  province = NULL,
  city = NULL,
  output = c("num", "chr"),
  grep = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{...}{valid numeric id or name.}

\item{output}{'num' means return numeric id, and 'chr' means return name id.}

\item{grep}{logical, if `TRUE` (default) will use regular expressions.}

\item{na.rm}{logical, if `TRUE` will remove `NA` value.}
}
\value{
character vector.
}
\description{
Helper functions to extract ID of CNmap by numeric-id or name-id.
}
\author{
Hou Yun
}
