% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-trans.R
\name{bd09_to_gcj02}
\alias{bd09_to_gcj02}
\alias{gcj02_to_bd09}
\alias{wgs84_to_gcj02}
\alias{gcj02_to_wgs84}
\title{Transform coordinate reference system}
\usage{
bd09_to_gcj02(lng, lat)

gcj02_to_bd09(lng, lat)

wgs84_to_gcj02(lng, lat)

gcj02_to_wgs84(lng, lat)
}
\arguments{
\item{lng}{numeric vector of longitude.}

\item{lat}{numeric vector of latitude.}
}
\value{
a new data frame of longitude and latitude.
}
\description{
A set of tools to convert points to points with another
coordinate reference system, and it's useful for Chinese users only.
}
