% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ad}
\alias{ad}
\title{the code and name of China's administrative divisions}
\format{
A data frame with 3212 rows and 2 variables:
\describe{
\item{ID}{character, the code of China's administrative divisions}
\item{name}{character, asscii code, the name of China's administrative divisions. The asscii code can be converted to Chinese character using rawToChar()}
...
}
}
\usage{
ad
}
\description{
the code and name of China's administrative divisions
}
\keyword{datasets}
